requirements.txt
Django==5.2.8
django-environ==0.12.0
djangorestframework==3.16.1
